# HHH-10835 The hash code calculation of the EntityKey should also include the hash code of the root entity name.