--- a/hibernate-core/src/main/java/org/hibernate/engine/spi/EntityKey.java
+++ b/hibernate-core/src/main/java/org/hibernate/engine/spi/EntityKey.java
@@ -56,6 +56,8 @@ public final class EntityKey implements Serializable {
 
 	private int generateHashCode() {
 		int result = 17;
+		final String rootEntityName = persister.getRootEntityName();
+		result = 37 * result + ( rootEntityName != null ? rootEntityName.hashCode() : 0 );
 		result = 37 * result + persister.getIdentifierType().getHashCode( identifier, persister.getFactory() );
 		return result;
 	}