https://hibernate.atlassian.net/browse/HHH-10835