@Sanne 2016-06-12
Could you explain why? We had this carefully optimised and during our tests I had seen that including the type didn't add much value over the relatively good hash spreads I was getting out of the ids. On the other hand, the simplified hash calculation had a nice impact on CPU (as the method became much simpler).

I only tested this thoroughly some ~2 years ago though, so my observations might be outdated, or not considering some other factor. Just very interested in how you figured you want to change this :)