@golovnin 2016-06-13
We have the following use case: 200 entity classes which use a sequence per entity class to create the primary key. An import job creates entities of this 200 entity classes in the same transaction and stores them. And because the tables and sequences for all of this entity classes were created at the same time, all of this sequences produce the same PK values in a transaction and we end up with linear searching in the HashMaps of the PersistentContext. To fix this problem we have considered two solutions:

- Fix Hibernate.
- Change sequences to produce different values.

And at the end we have decided to fix Hibernate.

Let me show you some numbers (here http://pastebin.com/YKSrKSHJ is the benchmark code):
```
# JMH 1.12 (released 73 days ago)
# VM version: JDK 1.8.0_92, VM 25.92-b14
# VM invoker: /Library/Java/JavaVirtualMachines/jdk1.8.0_92.jdk/Contents/Home/jre/bin/java
# VM options: <none>
# Warmup: 10 iterations, 3 s each
# Measurement: 10 iterations, 3 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: org.hibernate.engine.spi.EntityKeyBenchmark.createEntityKey

Before patch:

Result "createEntityKey":
  10.286 ±(99.9%) 0.151 ns/op [Average]
  (min, avg, max) = (9.544, 10.286, 13.170), stdev = 0.446
  CI (99.9%): [10.135, 10.438] (assumes normal distribution)


# Run complete. Total time: 00:10:08

Benchmark                           Mode  Cnt   Score   Error  Units
EntityKeyBenchmark.createEntityKey  avgt  100  10.286 ± 0.151  ns/op

After patch:

Result "createEntityKey":
  10.419 ±(99.9%) 0.186 ns/op [Average]
  (min, avg, max) = (9.815, 10.419, 12.766), stdev = 0.548
  CI (99.9%): [10.233, 10.604] (assumes normal distribution)


# Run complete. Total time: 00:10:09

Benchmark                           Mode  Cnt   Score   Error  Units
EntityKeyBenchmark.createEntityKey  avgt  100  10.419 ± 0.186  ns/op
```
As you can see there is no big difference between the current version and the patched one. Calculating the hash code of String is relatively cheap as the hash code is calculated only once (objects of the String class caches the calculated hash code).

If you decide to reject this PR (it is OK for me), then please remove the adding the constant value (`37*17`) to the calculated hash code of the identifier as it is useless and please add a comment to the method `EntityKey#generateHashCode()` why it does not use the root entity name to calculate the hash code.