@Sanne 2016-06-13
Thanks @golovnin for the excellent explanation and the JMH tests :+1: 

When I changed this we had also plans to separate the storage of entities within the persistence context based on their type. I still think that should be done to improve performance even further, but until that day comes (as it's a complex refactoring) I think we should apply your patch.